+++ src/main/resources/assets/opencomputers/loot/openos/bin/edit.lua
@@ -2,7 +2,7 @@ local event = require("event")
 local fs = require("filesystem")
 local keyboard = require("keyboard")
 local shell = require("shell")
-local term = require("term")
+local term = require("term") -- TODO use tty and cursor position instead of global area and gpu
 local text = require("text")
 local unicode = require("unicode")
