+++ src/main/resources/assets/opencomputers/loot/openos/bin/set.lua
@@ -1,5 +1,3 @@
-local text = require("text")
-
 local args = {...}
 
 if #args < 1 then
