+++ src/main/resources/assets/opencomputers/loot/openos/boot/93_term.lua
@@ -1,20 +1,17 @@
 local component = require("component")
 local computer = require("computer")
 local event = require("event")
-local term = require("term")
+local tty = require("tty")
 local process = require("process")
 
--- this should be the init level process
-process.info().data.window = term.internal.open()
-
 event.listen("gpu_bound", function(ename, gpu)
-  gpu=component.proxy(gpu)
-  term.bind(gpu)
+  gpu = component.proxy(gpu)
+  tty.bind(gpu)
   computer.pushSignal("term_available")
 end)
 
 local function components_changed(ename, address, type)
-  local window = term.internal.window()
+  local window = tty.window
   if not window then
     return
   end
@@ -47,7 +44,7 @@ local function components_changed(ename, address, type)
     window.keyboard = nil
   end
 
-  if (type == "screen" or type == "gpu") and not term.isAvailable() then
+  if (type == "screen" or type == "gpu") and not tty.isAvailable() then
     computer.pushSignal("term_unavailable")
   end
 end
