+++ src/main/resources/assets/opencomputers/loot/openos/init.lua
@@ -13,7 +13,7 @@ do
     invoke(addr, "close", handle)
     return load(buffer, "=" .. file, "bt", _G)
   end]], "=loadfile", "bt", _G)()
-  loadfile("/lib/tools/boot.lua")(loadfile)
+  loadfile("/opt/core/boot.lua")(loadfile)
 end
 
 while true do
