+++ src/main/resources/assets/opencomputers/loot/openos/lib/term.lua
@@ -1,606 +1,78 @@
+local tty = require("tty")
 local unicode = require("unicode")
-local event = require("event")
-local process = require("process")
-local kb = require("keyboard")
-local component = require("component")
 local computer = require("computer")
+local process = require("process")
+
+local kb = require("keyboard")
 local keys = kb.keys
 
-local term = {}
-term.internal = {}
+local term = setmetatable({internal={}}, {__index=tty})
 
 function term.internal.window()
   return process.info().data.window
 end
 
-local W = term.internal.window
-local local_env = {unicode=unicode,event=event,process=process,W=W,kb=kb}
-
-local gpu_intercept = {}
-local function update_viewport(window, width, height)
-  window = window or W()
-  local gpu = window.gpu
-  if not gpu then return end
-  if not gpu_intercept[gpu] then
-    gpu_intercept[gpu] = {} -- only override a gpu once
-    -- the gpu can change resolution before we get a chance to call events and handle screen_resized
-    -- unfortunately, we have to handle viewport changes by intercept
-    local setr, setv = gpu.setResolution, gpu.setViewport
-    gpu.setResolution = function(...)
-      gpu_intercept[gpu] = {}
-      return setr(...)
-    end
-    gpu.setViewport = function(...)
-      gpu_intercept[gpu] = {}
-      return setv(...)
-    end
+local function as_window(window, func, ...)
+  local data = process.info().data
+  if not data.window then
+    return func(...)
   end
-  if not width and not gpu_intercept[gpu][window] then
-    width, height = gpu.getViewport()
-  end
-  if width then
-    window:resize(width, height)
-    gpu_intercept[gpu][window] = true
-  end
-end
-
-local function resize(window, width, height)
-  window.w, window.h = width, height
+  local prev = rawget(data, "window")
+  data.window = window
+  local ret = table.pack(func(...))
+  data.window = prev
+  return table.unpack(ret, ret.n)
 end
 
 function term.internal.open(...)
   local dx, dy, w, h = ...
-  local window = {x=1,y=1,fullscreen=select("#",...)==0,dx=dx or 0,dy=dy or 0,w=w,h=h,blink=true,resize=resize}
-  event.listen("screen_resized", function(_,addr,w,h)
-    if term.isAvailable(window) and term.screen(window) == addr and window.fullscreen then
-      update_viewport(window, w, h)
+  local window = {fullscreen=select("#",...) == 0, blink = true}
+
+  -- support legacy code using direct manipulation of w and h
+  -- (e.g. wocchat) instead of using setViewport
+  setmetatable(window,
+  {
+    __index = function(tbl, key)
+      key = key == "w" and "width" or key == "h" and "height" or key
+      return rawget(tbl, key)
+    end,
+    __newindex = function(tbl, key, value)
+      key = key == "w" and "width" or key == "h" and "height" or key
+      return rawset(tbl, key, value)
     end
-  end)
+  })
+
+  -- first time we open a pty the current tty.window must become the process window
+  if not term.internal.window() then
+    local init_index = 2
+    while process.info(init_index) do
+      init_index = init_index + 1
+    end
+    process.info(init_index - 1).data.window = tty.window
+    tty.window = nil
+    setmetatable(tty,
+    {
+      __index = function(tbl, key)
+        if key == "window" then
+          return term.internal.window()
+        end
+      end
+    })
+  end
+
+  as_window(window, tty.setViewport, w, h, dx, dy, 1, 1)
   return window
 end
 
-function term.getViewport(window)
-  window = window or W()
-  update_viewport(window)
-  return window.w, window.h, window.dx, window.dy, window.x, window.y
-end
-
-function term.setViewport(w,h,dx,dy,x,y,window)
-  window = window or W()
-
-  dx,dy,x,y = dx or 0,dy or 0,x or 1,y or 1
-  if not w or not h then
-    local gw,gh = window.gpu.getViewport()
-    w,h = w or gw, h or gh
-  end
-
-  window.dx,window.dy,window.x,window.y,window.gw,window.gh = dx, dy, x, y, gw, gh
-  update_viewport(window, w, h)
-end
-
-function term.gpu(window)
-  window = window or W()
-  return window.gpu
-end
-
-function term.clear()
-  local w = W()
-  local gpu = w.gpu
-  if not gpu then return end
-  gpu.fill(1+w.dx,1+w.dy,w.w,w.h," ")
-  w.x,w.y=1,1
-end
-
-function term.isAvailable(w)
-  w = w or W()
-  return w and not not (w.gpu and w.gpu.getScreen())
-end
-
-function term.internal.pull(input, timeout, ...)
-  timeout = timeout or math.huge
-
-  local w = W()
-  local d, h, dx, dy, x, y = term.getViewport(w)
-  local out = (x<1 or x>d or y<1 or y>h)
-
-  if input and out then
-    input:move(0)
-    y = w.y
-    input:scroll()
-  end
-
-  x, y = w.x + dx, w.y + dy
-  local gpu = (input or not out) and w.gpu
-
-  local bgColor, bgIsPalette
-  local fgColor, fgIsPalette
-  local char_at_cursor
-  local blinking
-  if gpu then
-    bgColor, bgIsPalette = gpu.getBackground()
-    -- it can happen during a type of race condition when a screen is removed
-    if not bgColor then
-      return nil, "interrupted"
-    end
-
-    fgColor, fgIsPalette = gpu.getForeground()
-    char_at_cursor = gpu.get(x, y)
-
-    blinking = w.blink
-    if input then
-      blinking = input.blink
-    end
-  end
-
-  -- get the next event
-  local blinked = false
-  local done = false
-  local signal
-  while true do
-    if gpu then
-      if not blinked and not done then
-        gpu.setForeground(bgColor, bgIsPalette)
-        gpu.setBackground(fgColor, fgIsPalette)
-        gpu.set(x, y, char_at_cursor)
-        gpu.setForeground(fgColor, fgIsPalette)
-        gpu.setBackground(bgColor, bgIsPalette)
-        blinked = true
-      elseif blinked then
-        gpu.set(x, y, char_at_cursor)
-        blinked = false
-      end
-    end
-
-    if done then
-      return table.unpack(signal, 1, signal.n)
-    end
-
-    signal = table.pack(event.pull(math.min(.5, timeout), ...))
-    timeout = timeout - .5
-    done = signal.n > 1 or timeout < .5
-  end
-end
-
-function term.pull(...)
-  local args = table.pack(...)
-  local timeout = nil
-  if type(args[1]) == "number" then
-    timeout = table.remove(args, 1)
-    args.n = args.n - 1
-  end
-  return term.internal.pull(nil, timeout, table.unpack(args, 1, args.n))
-end
-
-function term.read(history,dobreak,hintHandler,pwchar,filter)
-  if not io.stdin.tty then return io.read() end
-  local ops = history or {}
-  ops.dobreak = ops.dobreak
-  if ops.dobreak==nil then ops.dobreak = dobreak end
-  ops.hintHandler = ops.hintHandler or hintHandler
-  ops.pwchar = ops.pwchar or pwchar
-  ops.filter = ops.filter or filter
-  return term.readKeyboard(ops)
-end
-
-function term.internal.split(input)
-  local data,index=input.data,input.index
-  local dlen = unicode.len(data)
-  index=math.max(0,math.min(index,dlen))
-  local tail=dlen-index
-  return unicode.sub(data,1,index),tail==0 and""or unicode.sub(data,-tail)
-end
-
-function term.internal.build_vertical_reader(input)
-  input.sy = 0
-  input.scroll = function(_)
-    _.sy = _.sy + term.internal.scroll(_.w)
-    _.w.y = math.min(_.w.y,_.w.h)
-  end
-  input.move = function(_,n)
-    local w=_.w
-    _.index = math.min(math.max(0,_.index+n),unicode.len(_.data))
-    local s1,s2 = term.internal.split(_)
-    s2 = unicode.sub(s2.." ",1,1)
-    local data_remaining = ("_"):rep(_.promptx-1)..s1..s2
-    w.y = _.prompty - _.sy
-    while true do
-      local wlen_remaining = unicode.wlen(data_remaining)
-      if wlen_remaining > w.w then
-        local line_cut = unicode.wtrunc(data_remaining, w.w+1)
-        data_remaining = unicode.sub(data_remaining,unicode.len(line_cut)+1)
-        w.y=w.y+1
-      else
-        w.x = wlen_remaining-unicode.wlen(s2)+1
-        break
-      end
-    end
-  end
-  input.clear_tail = function(_)
-    local win=_.w
-    local oi,w,h,dx,dy,ox,oy = _.index,term.getViewport(win)
-    _:move(math.huge)
-    _:move(-1)
-    local ex,ey=win.x,win.y
-    win.x,win.y,_.index=ox,oy,oi
-    x=oy==ey and ox or 1
-    win.gpu.fill(x+dx,ey+dy,w-x+1,1," ")
-  end
-  input.update = function(_,arg)
-    local w,cursor,suffix=_.w
-    local s1,s2=term.internal.split(_)
-    if type(arg) == "number" then
-      local ndata
-      if arg < 0 then if _.index<=0 then return end
-        _:move(-1)
-        ndata=unicode.sub(s1,1,-2)..s2
-      else if _.index>=unicode.len(_.data) then return end
-        s2=unicode.sub(s2,2)
-        ndata=s1..s2
-      end
-      suffix=s2
-      input:clear_tail()
-      _.data = ndata
-    else
-      _.data=s1..arg..s2
-      _.index=_.index+unicode.len(arg)
-      cursor,suffix=arg,s2
-    end
-    if cursor then _:draw(_.mask(cursor)) end
-    if suffix and suffix~="" then
-      local px,py,ps=w.x,w.y,_.sy
-      _:draw(_.mask(suffix))
-      w.x,w.y=px,py-(_.sy-ps)
-    end
-  end
-  input.clear = function(_)
-    _:move(-math.huge)
-    _:draw((" "):rep(unicode.wlen(_.data)))
-    _:move(-math.huge)
-    _.index=0
-    _.data=""
-  end
-  input.draw = function(_,text)
-    _.sy = _.sy + term.drawText(text,true)
-  end
-end
-
-function term.internal.read_history(history,input,change)
-  if not change then
-    if unicode.wlen(input.data) > 0 then
-      table.insert(history.list,1,input.data)
-      history.list[(tonumber(os.getenv("HISTSIZE")) or 10)+1]=nil
-      history.list[0]=nil
-    end
-  else
-    local ni = history.index + change
-    if ni >= 0 and ni <= #history.list then
-      history.list[history.index]=input.data
-      history.index = ni
-      input:clear()
-      input:update(history.list[ni])
-    end
-  end
-end
-
-function term.readKeyboard(ops)
-  checkArg(1,ops,"table")
-  local filter = ops.filter and function(i) return term.internal.filter(ops.filter,i) end or term.internal.nop
-  local pwchar = ops.pwchar and function(i) return term.internal.mask(ops.pwchar,i) end or term.internal.nop
-  local history,db,hints={list=ops,index=0},ops.dobreak,{handler=ops.hintHandler}
-  local w=W()
-  local draw=io.stdin.tty and term.drawText or term.internal.nop
-  local input={w=w,promptx=w.x,prompty=w.y,index=0,data="",mask=pwchar}
-  input.blink = ops.blink
-  if input.blink == nil then
-    input.blink = w.blink
-  end
-
-  -- two wrap types currently supported, vertical and hortizontal
-  if ops.nowrap then term.internal.build_horizontal_reader(input)
-  else               term.internal.build_vertical_reader(input)
-  end
-
-  while true do
-    local name, address, char, code = term.internal.pull(input)
-    if not term.isAvailable() then
-      return
-    end
-
-    -- we have to keep checking what kb is active in case it is switching during use
-    -- we could have multiple screens, each with keyboards active
-    local main_kb = term.keyboard(w)
-    local main_sc = term.screen(w)
-    local c
-    local backup_cache = hints.cache
-    if name == "interrupted" then
-      draw("^C\n",true)
-      return false
-    elseif address == main_kb or address == main_sc then
-      if name == "touch" or name == "drag" then
-        term.internal.onTouch(input,char,code)
-      elseif name == "clipboard" then
-        c = term.internal.clipboard(char)
-        hints.cache = nil
-      elseif name == "key_down" then
-        hints.cache = nil
-        local ctrl = kb.isControlDown(address)
-        if ctrl and code == keys.d then return
-        elseif code == keys.tab then
-          hints.cache = backup_cache
-          term.internal.tab(input,hints)
-        elseif (code == keys.enter or code == keys.numpadenter)
-                and filter(input) then
-          input:move(math.huge)
-          if db ~= false then
-            draw("\n")
-          end
-          term.internal.read_history(history,input)
-          return input.data .. "\n"
-        elseif code == keys.up     then term.internal.read_history(history, input,  1)
-        elseif code == keys.down   then term.internal.read_history(history, input, -1)
-        elseif code == keys.left   then input:move(ctrl and term.internal.ctrl_movement(input, -1) or -1)
-        elseif code == keys.right  then input:move(ctrl and term.internal.ctrl_movement(input,  1) or  1)
-        elseif code == keys.home   then input:move(-math.huge)
-        elseif code == keys["end"] then input:move( math.huge)
-        elseif code == keys.back   then c = -1
-        elseif code == keys.delete then c =  0
-        elseif ctrl and char == "w"then -- TODO: cut word
-        elseif char >= 32          then c = unicode.char(char)
-        else                            hints.cache = backup_cache -- ignored chars shouldn't clear hint cache
-        end
-      end
-      -- if we obtained something (c) to handle
-      if c then
-        input:update(c)
-      end
-    end
-  end
-end
-
--- cannot use term.write = io.write because io.write invokes metatable
-function term.write(value,wrap)
-  local stdout = io.output()
-  local stream = stdout and stdout.stream
-  local previous_wrap = stream.wrap
-  stream.wrap = wrap == nil and true or wrap
-  stdout:write(value)
-  stdout:flush()
-  stream.wrap = previous_wrap
-end
-
-function term.getCursor()
-  local w = W()
-  return w.x,w.y
-end
-
-function term.setCursor(x,y)
-  local w = W()
-  w.x,w.y=x,y
-end
-
-function term.drawText(value, wrap, window)
-  window = window or W()
-  if not window then return end
-  local gpu = window.gpu
-  if not gpu then return end
-  local w,h,dx,dy,x,y = term.getViewport(window)
-  local sy = 0
-  local vlen = #value
-  local index = 1
-  local cr_last,beeped = false,false
-  local function scroll(_sy,_y)
-    return _sy + term.internal.scroll(window,_y-h), math.min(_y,h)
-  end
-  local uptime = computer.uptime
-  local last_sleep = uptime()
-  while index <= vlen do
-    if uptime() - last_sleep > 4 then
-      os.sleep(0)
-      last_sleep = uptime()
-    end
-    local si,ei = value:find("[\t\r\n\a]", index)
-    si = si or vlen+1
-    if index==si then
-      local delim = value:sub(index, index)
-      if delim=="\t" then
-        x=((x-1)-((x-1)%8))+9
-      elseif delim=="\r" or (delim=="\n" and not cr_last) then
-        x,y=1,y+1
-        sy,y = scroll(sy,y)
-      elseif delim=="\a" and not beeped then
-        computer.beep()
-        beeped = true
-      end
-      cr_last = delim == "\r"
-    else
-      sy,y = scroll(sy,y)
-      si = si - 1
-      local next = value:sub(index, si)
-      local wlen_needed = unicode.wlen(next)
-      local slen = #next
-      local wlen_remaining = w - x + 1
-      local clean_end = ""
-      if wlen_remaining < wlen_needed then
-        next = unicode.wtrunc(next, wlen_remaining + 1)
-        wlen_needed = unicode.wlen(next)
-        clean_end = (" "):rep(wlen_remaining-wlen_needed)
-        if not wrap then
-          si = math.huge
-        end
-      end
-      gpu.set(x+dx,y+dy,next..clean_end)
-      x = x + wlen_needed
-      if wrap and slen ~= #next then
-        si = si - (slen - #next)
-        x = 1
-        y = y + 1
-      end
-    end
-    index = si + 1
-  end
-
-  window.x,window.y = x,y
-  return sy
-end
-
-function term.internal.scroll(w,n)
-  w = w or W()
-  local gpu,d,h,dx,dy,x,y = w.gpu,term.getViewport(w)
-  n = n or (y-h)
-  if n <= 0 then return 0 end
-  gpu.copy(dx+1,dy+n+1,d,h-n,0,-n)
-  gpu.fill(dx+1,dy+h-n+1,d,n," ")
-  return n
-end
-
-function term.internal.nop(...)
-  return ...
-end
-
-function term.setCursorBlink(enabled)
-  W().blink=enabled
-end
-
-function term.getCursorBlink()
-  return W().blink
-end
-
-function term.bind(gpu, window)
-  window = window or W()
-  window.gpu = gpu or window.gpu
-  window.keyboard = nil -- without a keyboard bound, always use the screen's main keyboard (1st)
-  if window.fullscreen then
-    term.setViewport(nil,nil,nil,nil,window.x,window.y,window)
-  end
-end
-
-function term.keyboard(window)
-  window = window or W() or {} -- this method needs to be safe even if there is no terminal window (e.g. no gpu)
-
-  if window.keyboard then
-    return window.keyboard
-  end
-
-  local system_keyboard = component.isAvailable("keyboard") and component.keyboard
-  system_keyboard = system_keyboard and system_keyboard.address or "no_system_keyboard"
-
-  local screen = term.screen(window)
-
-  if not screen then
-    -- no screen, no known keyboard, use system primary keyboard if any
-    return system_keyboard
-  end
-
-  -- if we are using a gpu bound to the primary scren, then use the primary keyboard
-  if component.isAvailable("screen") and component.screen.address == screen then
-    window.keyboard = system_keyboard
-  else
-    -- calling getKeyboards() on the screen is costly (time)
-    -- custom terminals should avoid designs that require
-    -- this on every key hit
-
-    -- this is expensive (slow!)
-    window.keyboard = component.invoke(screen, "getKeyboards")[1] or system_keyboard
-  end
-
-  return window.keyboard
-end
-
-function term.screen(window)
-  window = window or W()
-  local gpu = window.gpu
-  if not gpu then
-    return nil
-  end
-  return gpu.getScreen()
-end
-
-function --[[@delayloaded-start@]] term.scroll(number, window)
-  -- if zero scroll length is requested, do nothing
-  if number == 0 then return end
-  -- window is optional, default to current active terminal
-  window = window or W()
-  -- gpu works with global coordinates
-  local gpu,width,height,dx,dy,x,y = window.gpu,term.getViewport(w)
-
-  -- scroll request can be too large
-  local abs_number = math.abs(number)
-  if (abs_number >= height) then
-    term.clear()
-    return
-  end
-
-  -- box positions to shift
-  local box_height = height - abs_number
-  local top = 0
-  if number > 0 then
-    top = number -- (e.g. 1 scroll moves box at 2)
-  end
-
-  gpu.copy(dx + 1, dy + top + 1, width, box_height, 0, -number)
-
-  local fill_top = 0
-  if number > 0 then
-    fill_top = box_height
-  end
-
-  gpu.fill(dx + 1, dy + fill_top + 1, width, abs_number, " ")
-end --[[@delayloaded-end@]]
-
-function --[[@delayloaded-start@]] term.internal.ctrl_movement(input, dir)
-  local index, data = input.index, input.data
-
-  local function isEdge(char)
-    return char == "" or not not char:find("%s")
-  end
-
-  local last=dir<0 and 0 or unicode.len(data)
-  local start=index+dir+1
-  for i=start,last,dir do
-    local a,b = unicode.sub(data, i-1, i-1), unicode.sub(data, i, i)
-    if isEdge(a) and not isEdge(b) then return i-(index+1) end
-  end
-  return last - index
-end --[[@delayloaded-end@]]
-
-function --[[@delayloaded-start@]] term.internal.onTouch(input,gx,gy)
-  if input.data == "" then return end
-  input:move(-math.huge)
-  local w = W()
-  gx,gy=gx-w.dx,gy-w.dy
-  local x2,y2,d = input.w.x,input.w.y,input.w.w
-  local char_width_to_move = ((gy*d+gx)-(y2*d+x2))
-  if char_width_to_move <= 0 then return end
-  local total_wlen = unicode.wlen(input.data)
-  if char_width_to_move >= total_wlen then
-    input:move(math.huge)
-  else
-    local chars_to_move = unicode.wtrunc(input.data, char_width_to_move + 1)
-    input:move(unicode.len(chars_to_move))
-  end
-  -- fake white space can make the index off, redo adjustment for alignment
-  x2,y2,d = input.w.x,input.w.y,input.w.w
-  char_width_to_move = ((gy*d+gx)-(y2*d+x2))
-  if (char_width_to_move < 0) then
-    -- using char_width_to_move as a type of index is wrong, but large enough and helps to speed this up
-    local up_to_cursor = unicode.sub(input.data, input.index+char_width_to_move, input.index)
-    local full_wlen = unicode.wlen(up_to_cursor)
-    local without_tail = unicode.wtrunc(up_to_cursor, full_wlen + char_width_to_move + 1)
-    local chars_cut = unicode.len(up_to_cursor) - unicode.len(without_tail)
-    input:move(-chars_cut)
-  end
-end --[[@delayloaded-end@]]
-
-function --[[@delayloaded-start@]] term.internal.build_horizontal_reader(input)
-  term.internal.build_vertical_reader(input)
-  input.clear_tail = function(_)
-    local w,h,dx,dy,x,y = term.getViewport(_.w)
-    local s1,s2=term.internal.split(_)
+local function build_horizontal_reader(cursor)
+  cursor.clear_tail = function(_)
+    local w,h,dx,dy,x,y = tty.getViewport()
+    local s1,s2=tty.internal.split(_)
     local wlen = math.min(unicode.wlen(s2),w-x+1)
-    _.w.gpu.fill(x+dx,y+dy,wlen,1," ")
+    tty.gpu().fill(x+dx,y+dy,wlen,1," ")
   end
-  input.move = function(_,n)
-    local win = _.w
+  cursor.move = function(_,n)
+    local win = tty.window
     local a = _.index
     local b = math.max(0,math.min(unicode.len(_.data),_.index+n))
     _.index = b
@@ -609,13 +81,13 @@ function --[[@delayloaded-start@]] term.internal.build_horizontal_reader(input)
     win.x = win.x + wlen_moved * (n<0 and -1 or 1)
     _:scroll()
   end
-  input.draw = function(_,text)
-    term.drawText(text,false)
+  cursor.draw = function(_, text)
+    tty.drawText(text, true)
   end
-  input.scroll = function(_)
-    local win = _.w
+  cursor.scroll = function(_)
+    local win = tty.window
     local gpu,data,px,i = win.gpu,_.data,_.promptx,_.index
-    local w,h,dx,dy,x,y = term.getViewport(win)
+    local w,h,dx,dy,x,y = tty.getViewport()
     win.x = math.max(_.promptx, math.min(w, x))
     local len = unicode.len(data)
     local available,sx,sy,last = w-px+1,px+dx,y+dy,i==len
@@ -640,94 +112,118 @@ function --[[@delayloaded-start@]] term.internal.build_horizontal_reader(input)
       gpu.set(sx,sy,data)
     end
   end
-  input.clear = function(_)
-    local win = _.w
+  cursor.clear = function(_)
+    local win = tty.window
     local gpu,data,px=win.gpu,_.data,_.promptx
-    local w,h,dx,dy,x,y = term.getViewport(win)
+    local w,h,dx,dy,x,y = tty.getViewport()
     _.index,_.data,win.x=0,"",px
     gpu.fill(px+dx,y+dy,w-px+1-dx,1," ")
   end
-end --[[@delayloaded-end@]]
+end
 
-function --[[@delayloaded-start@]] term.clearLine(window)
-  window = window or W()
-  local w,h,dx,dy,x,y = term.getViewport(window)
-  window.gpu.fill(dx+1,dy+math.max(1,math.min(y,h)),w,1," ")
-  window.x=1
-end --[[@delayloaded-end@]]
-
-function --[[@delayloaded-start@]] term.internal.mask(mask,input)
-  if not mask then return input end
-  if type(mask) == "function" then return mask(input) end
-  return mask:rep(unicode.wlen(input))
-end --[[@delayloaded-end@]]
-
-function --[[@delayloaded-start@]] term.internal.filter(filter,input)
-  if not filter then return true
-  elseif type(filter) == "string" then return input.data:match(filter)
-  elseif filter(input.data) then return true
-  else require("computer").beep(2000, 0.1) end
-end --[[@delayloaded-end@]]
-
-function --[[@delayloaded-start@]] term.internal.tab(input,hints)
-  if not hints.handler then return end
-  local main_kb = term.keyboard()
-  -- term may not have a keyboard
-  -- in which case, we shouldn't be handling tab events
-  if not main_kb then
-    return
-  end
-  if not hints.cache then
-    local data = hints.handler
-    hints.handler = function(...)
-      if type(data) == "table" then
-        return data
-      else
-        return data(...) or {}
+local function inject_filter(handler, filter)
+  if filter then
+    if type(filter) == "string" then
+      local filter_text = filter
+      filter = function(text)
+        return text:match(filter_text)
       end
     end
-    hints.cache = hints.handler(input.data, input.index + 1)
-    hints.cache.i = -1
+
+    local mt =
+    {
+      __newindex = function(tbl, key, value)
+        if key == "key_down" then
+          local tty_key_down = value
+          value = function(handler, cursor, char, code)
+            if code == keys.enter or code == keys.numpadenter then
+              if not filter(cursor.data) then
+                computer.beep(2000, 0.1)
+                return false -- ignore
+              end
+            end
+            return tty_key_down(handler, cursor, char, code)
+          end
+        end
+        rawset(tbl, key, value)
+      end
+    }
+    setmetatable(handler, mt)
+  end
+end
+
+local function inject_mask(cursor, dobreak, pwchar)
+  if not pwchar and dobreak ~= false then
+    return
   end
 
-  local cache = hints.cache
-  local cache_size = #cache
-  
-  if cache_size == 1 and cache.i == 0 then
-    -- there was only one solution, and the user is asking for the next
-    hints.cache = hints.handler(cache[1], input.index + 1)
-    hints.cache.i = -1
-    cache = hints.cache
-    cache_size = #cache
-  end
-
-  local change = kb.isShiftDown(main_kb) and -1 or 1
-  cache.i = (cache.i + change) % math.max(#cache, 1)
-  local next = cache[cache.i + 1]
-  if next then
-    local tail = unicode.len(input.data) - input.index
-    input:clear()
-    input:update(next)
-    input:move(-tail)
-  end
-end --[[@delayloaded-end@]]
-
-function --[[@delayloaded-start@]] term.getGlobalArea(window)
-  local w,h,dx,dy = term.getViewport(window)
-  return dx+1,dy+1,w,h
-end --[[@delayloaded-end@]]
-
-function --[[@delayloaded-start@]] term.internal.clipboard(char)
-  local first_line, end_index = char:find("\13?\10")
-  if first_line then
-    local after = char:sub(end_index + 1)
-    if after ~= "" then
-      require("computer").pushSignal("key_down", term.keyboard(), 13, 28)
-      require("computer").pushSignal("clipboard", term.keyboard(), after)
+  if pwchar then
+    if type(pwchar) == "string" then
+      local pwchar_text = pwchar
+      pwchar = function(text)
+        return text:gsub(".", pwchar_text)
+      end
     end
-    char = char:sub(1, first_line - 1)
   end
-  return char
-end --[[@delayloaded-end@]]
 
-return term, local_env
+  local cursor_draw = cursor.draw
+  cursor.draw = function(cursor, text)
+    local pre, newline = text:match("(.-)(\n?)$")
+    if dobreak == false then
+      newline = ""
+    end
+    if pwchar then
+      pre = pwchar(pre)
+    end
+    return cursor_draw(cursor, pre .. newline)
+  end
+end
+
+function term.read(history, dobreak, hint, pwchar, filter)
+  if not io.stdin.tty then
+    return io.read()
+  end
+  local handler = history or {}
+  handler.hint = handler.hint or hint
+
+  local cursor = tty.internal.build_vertical_reader()
+  if handler.nowrap then
+    build_horizontal_reader(cursor)
+  end
+
+  inject_filter(handler, filter)
+  inject_mask(cursor, dobreak, pwchar)
+  -- todo, make blinking work from here
+  -- handler.blink or w.blink
+
+  return tty.read(handler, cursor)
+end
+
+function term.getGlobalArea(window)
+  local w,h,dx,dy = as_window(window, tty.getViewport)
+  return dx+1,dy+1,w,h
+end
+
+function term.clearLine(window)
+  window = window or tty.window
+  local w,h,dx,dy,x,y = as_window(window, tty.getViewport)
+  window.gpu.fill(dx+1,dy+math.max(1,math.min(y,h)),w,1," ")
+  window.x=1
+end
+
+function term.pull(...)
+  local args = table.pack(...)
+  local timeout = nil
+  if type(args[1]) == "number" then
+    timeout = table.remove(args, 1)
+    args.n = args.n - 1
+  end
+  return tty.pull(nil, timeout, table.unpack(args, 1, args.n))
+end
+
+function term.bind(gpu, window)
+  return as_window(window, tty.bind, gpu)
+end
+
+return term
+
