+++ src/main/resources/assets/opencomputers/loot/openos/opt/core/devfs/adapters/internet.lua
