+++ src/main/resources/assets/opencomputers/loot/openos/opt/core/device_labeling.lua
