+++ src/main/resources/assets/opencomputers/loot/openos/opt/core/install_utils.lua
@@ -31,6 +31,7 @@ local function select_prompt(devs, prompt)
         choice = devs[number]
       else
         io.write("Invalid input, please try again: ")
+        os.sleep(0)
       end
     end
   end
