+++ src/main/resources/assets/opencomputers/loot/openos/opt/core/ro_wrapper.lua
